// src/app/model/api-model/ApiItem.ts
export class ApiItem{

  id: string;
  title: string;
  src_endpoint_id: string;
  top_priority: number;
  html_src_output: string;
  html_esb_ctrl: string;
  html_dst_input: string;
  html_dst_reply: string;

  html_any_xsmall: string;
  html_any_small: string;
  html_any_medium: string;
  html_any_big: string;

  searching_tag1: any;
  searching_tag2: any;
  searching_tag3: any;
  searching_tag4: any;
  searching_tag5: any;
  searching_tag6: any;
  searching_tag7: any;
  searching_tag8: any;
  searching_tag9: any;
  searching_tag10: any;
  searching_tag11: any;
  searching_tag12: any;
  searching_tag13: any;
  searching_tag14: any;
  searching_tag15: any;
  searching_tag16: any;
  searching_tag17: any;
  searching_tag18: any;
  searching_tag19: any;
  searching_tag20: any;
}